Move srl to its own package
