+++ src/main/scala/org/clulab/wm/eidos/serialization/jsonld/JLDSerializer.scala
@@ -22,8 +22,9 @@ import org.clulab.wm.eidos.document.attachments.DctDocumentAttachment
 import org.clulab.wm.eidos.document.attachments.LocationDocumentAttachment
 import org.clulab.wm.eidos.document.attachments.RelevanceDocumentAttachment
 import org.clulab.wm.eidos.document.attachments.TitleDocumentAttachment
-import org.clulab.wm.eidos.groundings.grounders.{AdjectiveGrounding, PredicateTuple}
-import org.clulab.wm.eidos.groundings.{OntologyGrounding, PredicateGrounding, OntologyNodeGrounding}
+import org.clulab.wm.eidos.groundings.grounders.AdjectiveGrounding
+import org.clulab.wm.eidos.groundings.grounders.srl.PredicateTuple
+import org.clulab.wm.eidos.groundings.{OntologyGrounding, OntologyNodeGrounding, PredicateGrounding}
 import org.clulab.wm.eidos.mentions.{EidosCrossSentenceEventMention, EidosCrossSentenceMention, EidosEventMention, EidosMention, EidosTextBoundMention}
 import org.clulab.wm.eidos.utils.Unordered
 import org.clulab.wm.eidos.utils.Unordered.OrderingOrElseBy
